Remove favoritecard directory and files, moved components and accompanying files into new directories, and fixed imports to direct to the new locations
