+++ src/App/App.css
@@ -3,14 +3,14 @@
 }
 
 .App {
-  background: url('./images/starry-night.jpg');
+  background: url('../images/starry-night.jpg');
   height: 100vh;
   width: 100vw;
   z-index: -1000;
 }
 
 .cockpit {
-  background: url('./images/cockpit.png') top center no-repeat;
+  background: url('../images/cockpit.png') top center no-repeat;
   background-size: cover;
   height: 100vh;
   width: 100vw;
