+++ src/App/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import Favorites from './Favorites/Favorites'
-import Landing from './Landing';
-import Movies from './Movies';
+import Favorites from '../Favorites/Favorites'
+import Landing from '../Landing/Landing';
+import Movies from '../Movies/Movies';
 import './App.css';
 import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 
@@ -51,6 +51,9 @@ class App extends Component {
               </Route>
               <Route path='/movies'>
                 <Movies movies={this.state.movies}/>
+              </Route>
+              <Route path='/favorites'>
+                <Favorites />
               </Route>
             </Switch>
           </Router>
