--- src/FavoriteCard/FavoriteCard.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import './FavoriteCard.css';
-
-const FavoriteCard = () => {
-  return (
-    <div className='favorite-card'>
-
-    </div>
-  )
-}
-
-export default FavoriteCard;
